moving utils.lua to more suited utils folder
