+++ ftplugin/groovy.lua
@@ -13,7 +13,7 @@ vim.api.nvim_create_autocmd('BufWritePre', {
   group = vim.api.nvim_create_augroup('onSaveEventGroovy', { clear = true }),
   pattern = { "Jenkinsfile", "*.groovy" }, -- supported format on save
   callback = function()
-    -- require("custom.utils").format_basic_nvim()
+    -- require("utils.utils").format_basic_nvim()
     require("jenkinsfile_linter").validate()
   end,
 })
